chore: change the route of blog
